+++ pkg/ruler/group/compat.go
@@ -1,4 +1,4 @@
-package rulegroup
+package group
 
 import (
 	time "time"
