+++ pkg/ruler/group/rulegroup.go
@@ -1,4 +1,4 @@
-package rulegroup
+package group
 
 import (
 	"context"
