+++ pkg/ruler/group/rulegroup.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-package rulegroup;
+package group;
 
-option go_package = "rulegroup";
+option go_package = "group";
 
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 import "google/protobuf/duration.proto";
